@@ -9,6 +9,8 @@ import { EffectsModule } from '@ngrx/effects';
 import { UIEffects } from './store/ui/effects';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
 import { dashboardReducer } from './store';
+import { ProfileService } from './services/profile.service';
+import { WebsocketService } from './services/websocket.service';
 
 
 @NgModule({
@@ -22,7 +24,11 @@ import { dashboardReducer } from './store';
         UiModule,
         StoreModule.forFeature("dashboard", dashboardReducer),
         EffectsModule.forFeature([UIEffects]),
-        StoreDevtoolsModule.instrument({ maxAge: 25, logOnly: !isDevMode() })
+        StoreDevtoolsModule.instrument({ maxAge: 25, logOnly: !isDevMode() , connectInZone: true})
+    ],
+    providers: [
+        ProfileService,
+        WebsocketService
     ]
 })
 export class DashboardModule { }
