@@ -2,6 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ProfileComponent } from './profile.component';
 import { RouterModule, Routes } from '@angular/router';
+import { ReactiveFormsModule } from '@angular/forms';
+import { TextFormComponent } from './components/text-form/text-form.component';
 
 
 const routes: Routes = [
@@ -10,11 +12,13 @@ const routes: Routes = [
 
 @NgModule({
     declarations: [
-        ProfileComponent
+        ProfileComponent,
+        TextFormComponent
     ],
     imports: [
         CommonModule,
-        RouterModule.forChild(routes)
+        RouterModule.forChild(routes),
+        ReactiveFormsModule
     ]
 })
 export class ProfileModule { }
